Modifying new scheduler to limit who can play weeks 1-6.
